[SQUASH]: Rename "common/sfserr" to "errors"
