+++ src/common/exception.h
@@ -26,7 +26,7 @@
 #include <exception>
 #include <string>
 
-#include "common/sfserr/saunafs_error_codes.h"
+#include "errors/saunafs_error_codes.h"
 
 class Exception : public std::exception {
 public:
