+++ src/common/massert.h
@@ -27,7 +27,7 @@
 #include <stdlib.h>
 #include <sys/syslog.h>
 #include "common/slogger.h"
-#include "common/sfserr/sfserr.h"
+#include "errors/sfserr.h"
 
 #ifdef THROW_INSTEAD_OF_ABORT
 #  include <stdexcept>
