+++ src/mount/client/client_error_code.cc
@@ -20,8 +20,8 @@
 
 #include <string>
 #include "common/platform.h"
-#include "common/sfserr/saunafs_error_codes.h"
-#include "common/sfserr/sfserr.h"
+#include "errors/saunafs_error_codes.h"
+#include "errors/sfserr.h"
 
 #include "client_error_code.h"
