+++ src/tools/set_eattr.cc
@@ -25,8 +25,8 @@
 #include <stdlib.h>
 
 #include "common/datapack.h"
-#include "common/sfserr/saunafs_error_codes.h"
-#include "common/sfserr/sfserr.h"
+#include "errors/saunafs_error_codes.h"
+#include "errors/sfserr.h"
 #include "tools/tools_commands.h"
 #include "tools/tools_common_functions.h"
